LayoutTests/http/tests/websocket/tests/hybi/close-code-and-reason.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test CloseEvent code and reason property.");

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

var closeEvent;
var codeNoStatusRcvd = 1005;
var codeAbnormalClosure = 1006;
var emptyString = "";

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/hybi/echo");

ws.onopen = function(event)
{
    testFailed("ws.onopen() was called. (message = \"" + event.data + "\")");
};

ws.onclose = function(event)
{
    debug("ws.onclose() was called.");
    closeEvent = event;
    shouldBeFalse("closeEvent.wasClean");
    shouldBe("closeEvent.code", "codeAbnormalClosure");
    shouldBe("closeEvent.reason", "emptyString");
};

ws.close();

var testId = 0;
var testNum = 3;
var sendData = [
    "-", // request close frame without code and reason
    "--", // request close frame with invalid body which size is 1
    "65535 good bye", // request close frame with specified code and reason
];
var expectedCode = [
    codeNoStatusRcvd,
    codeNoStatusRcvd,
    65535,
];
var expectedReason = [
    "",
    "",
    "good bye",
];

WebSocketTest = function() {
    this.ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/hybi/close-code-and-reason");
    this.ws.onopen = this.onopen;
    this.ws.onmessage = this.onmessage;
    this.ws.onclose = this.onclose;
};

WebSocketTest.prototype.onopen = function()
{
    debug("WebSocketTest.onopen() was called with testId = " + testId + ".");
    this.send(sendData[testId]);
};

WebSocketTest.prototype.onmessage = function(event)
{
    testFailed("WebSocketTest.onmessage() was called. (message = \"" + event.data + "\")");
};

WebSocketTest.prototype.onclose = function(event)
{
    closeEvent = event;
    debug("WebSocketTest.onclose() was called with testId = " + testId + ".");
    shouldBeTrue("closeEvent.wasClean");
    shouldBe("closeEvent.code", "expectedCode[testId]");
    shouldBe("closeEvent.reason", "expectedReason[testId]");
    closeEvent.code = 0;
    closeEvent.reason = "readonly";
    shouldBe("closeEvent.code", "expectedCode[testId]");
    shouldBe("closeEvent.reason", "expectedReason[testId]");
    testId++;
    if (testId < testNum)
        test = new WebSocketTest();
    else
        finishJSTest();
};

var test = new WebSocketTest();

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
